Рабочая версия 2.0 Класс CourseAnalyser вынесен в отдельный модуль
