+++ .gitignore
@@ -129,3 +129,5 @@ dmypy.json
 .pyre/
 
 *.xls
+*.xlsx
+*.txt
